`platform_pki_2_electric_boogaloo` Windows implementation: re-use `mod.rs`
- Rename to `pki_system.rs` for consistency
- Rename `PkiSystem` -> `PlatformPkiSystem`
- Rename `PkiConfig` -> `PkiScwsConfig` and make it optional when initializing `PlatformPkiSystem`
  (as this config is only needed for SCWS)
- Rename `find_certificate` -> `open_certificate` (since "finding" as certificate is expected to be
  done using `list_user_certificates`)
- Remove `is_available` (as `PkiSystem::init` error should be used instead)
- Move `sign_message` and `decrypt_message` to `pki_private_key.rs`
- Rename `list_intermediate_certificates` -> `list_intermediate_x509_certificates`
- Rename `list_trusted_root_certificate_anchors` -> `list_x509_trust_anchors`
